.papelera/UTILITIES-fallido-20250612/batman-incorporated/requirements.txt
# Batman Incorporated Dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
gitpython>=3.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
